move _calculate_mid_points from _utils.py to geohash
