+++ dask_geopandas/geohash.py
@@ -10,7 +10,6 @@ The vectorized implementation for quantization and bit interleaving is in turn b
 """
 import numpy as np
 import pandas as pd
-from ._utils import _calculate_mid_points
 
 
 # Implementation based on deprecated https://pypi.org/project/neathgeohash/#description
@@ -45,6 +44,28 @@ def _geohash(gdf, string, p):
     geohash = encode_geohash(coords, string, p)
 
     return pd.Series(geohash, index=gdf.index, name="geohash")
+
+
+def _calculate_mid_points(bounds):
+    """
+    Calculate middle points based on the geometry bounds
+
+    Parameters
+    ----------
+    bounds : array_like
+        array containing xmin, ymin, xmax, ymax
+
+    Returns
+    ---------
+    x_mids : mid points of x values
+    y_mids : mid points of y values
+    """
+
+    # Calculate mid points for x and y bound coords
+    x_mids = (bounds[:, 0] + bounds[:, 2]) / 2.0
+    y_mids = (bounds[:, 1] + bounds[:, 3]) / 2.0
+
+    return x_mids, y_mids
 
 
 def encode_geohash(coords, string, p):
